33 relacionar la eliminacion de usuarios (#34)
* eliminar las sessiones de los usuarios al eliminarlo

* eliminar las sessiones de los usuarios al eliminarlo

* solucionar error de credenciales duplicadas

* eliminar sesiones de un usuario antes de eliminarlo

* comentarios

+++ app/users/application/DeleteUser.py
@@ -3,13 +3,23 @@ from flask import abort
 from app.users.domain.UserId import UserId
 from app.users.domain.UserStructure import UserStructure
 from app.users.interface.FirestoreRepository import UserRepository
+from app.sessions.interface.FirestoreRepository import SessionRepository
 
 
 class DeleteUser(UserStructure):
     def execute(self, fill_id):
+        # validate id
         user_id = UserId(fill_id)
         if user_id.is_valid() is False:
             abort(400)
 
-        repo = UserRepository()
-        return repo.deleteUser(user_id.value)
+        # delete session
+        session_repository = SessionRepository()
+        result = session_repository.delete_by_user_id(user_id.value)
+
+        if result:
+            # delete user
+            repo = UserRepository()
+            return repo.deleteUser(user_id.value)
+        else:
+            abort(400)
